Add non-nested solution to automatic keys
The Element sets its key in this priority:

- To `props.key`
- To the argument passed as `toReact(id)`

When the children of a Faux Element are a list,
a unique key is automatically generated. It can
be changed by setting `el.key`. This is not the
same as `props.key`, since `props.key` is always
in use, and `this.key` is not to be accessed by
the user normally.